Extract corruption detection into dedicated microcrate
