+++ copybook-codec/Cargo.toml
@@ -31,6 +31,7 @@ copybook-rdw.workspace = true
 copybook-record-io.workspace = true
 copybook-determinism.workspace = true
 copybook-error.workspace = true
+copybook-corruption.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
